Refactored MachineCard into it's own file
