+++ app/screens/LocationDetails.js
@@ -10,7 +10,6 @@ import {
   Share,
   StyleSheet,
   View,
-  Image,
 } from "react-native";
 import Mapbox from "@rnmapbox/maps";
 import openMap from "react-native-open-maps";
@@ -40,20 +39,13 @@ import {
 import * as WebBrowser from "expo-web-browser";
 import { useNavigation, useTheme } from "@react-navigation/native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
+import MachineCard from "../components/MachineCard";
 
 Mapbox.setAccessToken(process.env.MAPBOX_PUBLIC);
 
 let deviceWidth = Dimensions.get("window").width;
 
 const moment = require("moment");
-
-// TODO: pull from settings after adding setting option
-const showInsiderConnectedBadge = true;
-
-const insiderConnectedImage = {
-  dark: require("../assets/images/Insider_Connected_Vertical_RED_and_WHT_300_dpi.png"),
-  light: require("../assets/images/Insider_Connected_Vertical_standard_300_dpi.png"),
-};
 
 const LocationDetails = (props) => {
   const { route } = props;
@@ -109,78 +101,6 @@ const LocationDetails = (props) => {
     const positionY = event.nativeEvent.contentOffset.y;
     setShowScrollToTop(positionY > 150);
   };
-
-  const Title = ({ machine }) => (
-    <Text>
-      <Text style={s.machineName}>{machine.name}</Text>
-      {machine.year ? (
-        <Text style={[s.fontSize20, s.pink1, s.mediumFont]}>{` (${
-          machine.manufacturer && machine.manufacturer + ", "
-        }${machine.year})`}</Text>
-      ) : null}
-    </Text>
-  );
-
-  const MachineCard = ({ pressed, machine }) => (
-    <View>
-      <View
-        style={[
-          s.machineListContainer,
-          pressed ? s.pressed : s.notPressed,
-          { flexDirection: "row" },
-        ]}
-      >
-        <View style={{ flex: 1 }}>
-          <Title machine={machine} />
-          <View
-            style={{
-              flexDirection: "row",
-              alignItems: "center",
-              marginTop: 6,
-            }}
-          >
-            <MaterialCommunityIcons
-              name={
-                machine.created_at != machine.updated_at
-                  ? "clock-time-four-outline"
-                  : "clock-time-three-outline"
-              }
-              style={s.metaIcon}
-            />
-            <Text style={s.updated}>
-              {machine.created_at != machine.updated_at
-                ? `Updated: ${moment(machine.updated_at).format("MMM DD, YYYY")}`
-                : `Added: ${moment(machine.created_at).format("MMM DD, YYYY")}`}
-            </Text>
-          </View>
-        </View>
-        {showInsiderConnectedBadge && machine.ic_enabled && (
-          <View
-            style={{
-              flexDirection: "row",
-              justifyContent: "flex-end",
-              alignItems: "center",
-            }}
-          >
-            <Image
-              source={
-                theme.theme === "dark"
-                  ? insiderConnectedImage.dark
-                  : insiderConnectedImage.light
-              }
-              style={{
-                width: 50,
-                height: 50,
-                // Translation account for the space on the right and small TM so the card feels balanced
-                transform: [{ translateX: 10 }],
-              }}
-              resizeMode="contain"
-            />
-          </View>
-        )}
-      </View>
-    </View>
-  );
 
   const handleConfirmPress = (id, loggedIn) => {
     if (loggedIn) {
@@ -752,7 +672,7 @@ const LocationDetails = (props) => {
                   }}
                 >
                   {({ pressed }) => (
-                    <MachineCard pressed={pressed} machine={machine} />
+                    <MachineCard pressed={pressed} machine={machine} s={s} />
                   )}
                 </Pressable>
               ))}
